# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A comprehensive development toolkit for Odoo environments with profiling, monitoring, and optimization features.",
        "title": "OdooDevTools API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://example.com/support"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/environments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all environments for the current tenant",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "environments"
                ],
                "summary": "List environments",
                "parameters": [
                    {
                        "enum": [
                            "development",
                            "staging",
                            "production"
                        ],
                        "type": "string",
                        "description": "Filter by type",
                        "name": "env_type",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "active",
                            "inactive",
                            "maintenance"
                        ],
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 20, max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.EnvironmentListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Register a new Odoo environment under the current tenant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "environments"
                ],
                "summary": "Create environment",
                "parameters": [
                    {
                        "description": "Environment details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateEnvironmentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.EnvironmentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/environments/{env_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get details of a specific environment",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "environments"
                ],
                "summary": "Get environment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Environment ID (UUID)",
                        "name": "env_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.EnvironmentResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permanently delete an environment and its associated data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "environments"
                ],
                "summary": "Delete environment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Environment ID (UUID)",
                        "name": "env_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Partially update an environment's configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "environments"
                ],
                "summary": "Update environment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Environment ID (UUID)",
                        "name": "env_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateEnvironmentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.EnvironmentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
        "/environments/{env_id}/agent": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Assign an agent ID to an environment, setting its status to connected",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "environments"
                ],
                "summary": "Register agent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Environment ID (UUID)",
                        "name": "env_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Agent details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RegisterAgentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.EnvironmentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/api.ErrorCode"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ErrorDetail"
                    }
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
        "api.ErrorCode": {
            "type": "string",
            "enum": [
                "VALIDATION_ERROR",
                "UNAUTHORIZED",
                "FORBIDDEN",
                "NOT_FOUND",
                "CONFLICT",
                "UNPROCESSABLE_ENTITY",
                "RATE_LIMITED",
                "BAD_REQUEST",
                "PAYLOAD_TOO_LARGE",
                "INTERNAL_ERROR",
                "SERVICE_UNAVAILABLE",
                "TIMEOUT",
                "DATABASE_ERROR",
                "TENANT_NOT_FOUND",
                "ENVIRONMENT_NOT_FOUND",
                "AGENT_NOT_CONNECTED",
                "AGENT_NOT_FOUND",
                "QUOTA_EXCEEDED",
                "PLAN_LIMIT_REACHED",
                "INVALID_API_KEY",
                "EXPIRED_API_KEY",
                "INVALID_TOKEN",
                "EXPIRED_TOKEN",
                "INVALID_CREDENTIALS",
                "EMAIL_ALREADY_EXISTS",
                "SLUG_ALREADY_EXISTS",
                "INVALID_SCOPE",
                "SCHEMA_SNAPSHOT_NOT_FOUND",
                "MIGRATION_SCAN_NOT_FOUND",
                "ERROR_GROUP_NOT_FOUND",
                "ALERT_RULE_NOT_FOUND",
                "ALERT_CHANNEL_NOT_FOUND",
                "PROFILER_RECORDING_NOT_FOUND",
                "ANON_PROFILE_NOT_FOUND",
                "ANON_JOB_NOT_FOUND",
                "INVALID_ODOO_VERSION",
                "INVALID_DOMAIN"
            ],
            "x-enum-comments": {
                "ErrCodeExpiredAPIKey": "nolint:gosec // not a credential",
                "ErrCodeInvalidAPIKey": "nolint:gosec // not a credential",
                "ErrCodeInvalidCredentials": "nolint:gosec // not a credential"
            },
            "x-enum-descriptions": [
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "nolint:gosec // not a credential",
                "nolint:gosec // not a credential",
                "",
                "",
                "nolint:gosec // not a credential",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                "",
                ""
            ],
            "x-enum-varnames": [
                "ErrCodeValidation",
                "ErrCodeUnauthorized",
                "ErrCodeForbidden",
                "ErrCodeNotFound",
                "ErrCodeConflict",
                "ErrCodeUnprocessable",
                "ErrCodeRateLimited",
                "ErrCodeBadRequest",
                "ErrCodePayloadTooLarge",
                "ErrCodeInternal",
                "ErrCodeUnavailable",
                "ErrCodeTimeout",
                "ErrCodeDatabase",
                "ErrCodeTenantNotFound",
                "ErrCodeEnvironmentNotFound",
                "ErrCodeAgentNotConnected",
                "ErrCodeAgentNotFound",
                "ErrCodeQuotaExceeded",
                "ErrCodePlanLimitReached",
                "ErrCodeInvalidAPIKey",
                "ErrCodeExpiredAPIKey",
                "ErrCodeInvalidToken",
                "ErrCodeExpiredToken",
                "ErrCodeInvalidCredentials",
                "ErrCodeEmailAlreadyExists",
                "ErrCodeSlugAlreadyExists",
                "ErrCodeInvalidScope",
                "ErrCodeSchemaSnapshotNotFound",
                "ErrCodeMigrationScanNotFound",
                "ErrCodeErrorGroupNotFound",
                "ErrCodeAlertRuleNotFound",
                "ErrCodeAlertChannelNotFound",
                "ErrCodeProfilerRecordingNotFound",
                "ErrCodeAnonProfileNotFound",
                "ErrCodeAnonJobNotFound",
                "ErrCodeInvalidOdooVersion",
                "ErrCodeInvalidDomain"
            ]
        },
        "api.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "value": {}
            }
        },
...
        "dto.CreateEnvironmentRequest": {
            "type": "object",
            "required": [
                "db_name",
                "env_type",
                "name",
                "odoo_url"
            ],
            "properties": {
                "db_name": {
                    "type": "string",
                    "maxLength": 63,
                    "minLength": 1
                },
                "env_type": {
                    "type": "string",
                    "enum": [
                        "development",
                        "staging",
                        "production"
                    ]
                },
                "feature_flags": {
                    "description": "FeatureFlags holds arbitrary JSON configuration\n@Description Feature flags as JSON objec\n@example {\"enable_profiling\": true, \"log_level\": \"info\"}",
                    "type": "object"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "odoo_url": {
                    "type": "string"
                },
                "odoo_version": {
                    "type": "string",
                    "enum": [
                        "14.0",
                        "15.0",
                        "16.0",
                        "17.0",
                        "18.0"
                    ]
                }
            }
        },
        "dto.EnvironmentListResponse": {
            "type": "object",
            "properties": {
                "environments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.EnvironmentResponse"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dto.EnvironmentResponse": {
            "type": "object",
            "properties": {
                "agent_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "db_name": {
                    "type": "string"
                },
                "env_type": {
                    "type": "string"
                },
                "feature_flags": {
                    "type": "object"
                },
                "id": {
                    "type": "string"
                },
                "last_sync": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "odoo_url": {
                    "type": "string"
                },
                "odoo_version": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "dto.RegisterAgentRequest": {
            "type": "object",
            "required": [
                "agent_id"
            ],
            "properties": {
                "agent_id": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                }
            }
        },
...
        "dto.UpdateEnvironmentRequest": {
            "type": "object",
            "properties": {
                "db_name": {
                    "type": "string",
                    "maxLength": 63,
                    "minLength": 1
                },
                "env_type": {
                    "type": "string",
                    "enum": [
                        "development",
                        "staging",
                        "production"
                    ]
                },
                "feature_flags": {
                    "description": "@example {\"enable_profiling\": true, \"log_level\": \"info\"}",
                    "type": "object"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "odoo_url": {
                    "type": "string"
                },
                "odoo_version": {
                    "type": "string",
                    "enum": [
                        "14.0",
                        "15.0",
                        "16.0",
                        "17.0",
                        "18.0"
                    ]
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "active",
                        "inactive",
                        "maintenance"
                    ]
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Type \"ApiKey\" followed by a space and your API key (e.g. \"ApiKey odt_xxx...\").",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}